+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## HEAD
 
+* Upon confirmation redirect to original path.
+* Remove concept of after_signup route.
 * Enable OTP to be given in i18n mailer subject.